refactor(navigation): 收敛 ModLoaderSelector 正文页结构
- 将 ModLoaderSelectorPage 改为唯一的正文 Drill 页面，去掉页内 Header 壳层
- 删除 ModLoaderSelectorContentPage 与 ModLoaderSelectorContentView 两层中转结构
- ResourceDownload 的 secondary 导航直接指向 ModLoaderSelectorPage，并保持主项目构建通过
