--- XianYuLauncher/Features/ModLoaderSelector/Views/ModLoaderSelectorContentPage.xaml
@@ -1,19 +1,0 @@
-<Page
-    x:Class="XianYuLauncher.Features.ModLoaderSelector.Views.ModLoaderSelectorContentPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:controls="using:XianYuLauncher.Features.ModLoaderSelector.Controls"
-    xmlns:mediaAnimation="using:Microsoft.UI.Xaml.Media.Animation"
-    mc:Ignorable="d"
-    HorizontalAlignment="Stretch">
-
-        <Grid x:Name="ContentArea"
-          Background="Transparent"
-          mediaAnimation:EntranceNavigationTransitionInfo.IsTargetElement="True">
-        <controls:ModLoaderSelectorContentView Margin="24,0,24,24"
-                                               ViewModel="{x:Bind ViewModel}"
-                                               ShowVersionSelector="False" />
-    </Grid>
-</Page>
